MultiMoveNetCreate

Creates dynamic, directed, multiplex movement networks from capture-mark-recapture datasets using information on the capture locations and times of individuals.

MultiMoveNetCreate(
  data,
  intwindow,
  mindate,
  maxdate,
  netwindow,
  overlap,
  nextonly = FALSE,
  index = FALSE
)

Arguments

data
A 6 column dataframe with columns for the ID of the captured individual, the location of its capture (a name or number), the x coordinate of its capture location, the y coordinate of the capture location, the date of capture and information to define multiplex network layers. The "Layers" column can consist of any unique identifiers (e.g. if layers representing movements by males and females are used then they could be represented by "M" and "F" or 1 and 2). If the user wants a layer per individual then the "Layers" column can simply be a copy of the individual ID column.

intwindow
The maximum period of time (in days) between the capture of an individual at two different locations for it to be added as an edge to the movemenet network (i.e. if intwindow = 10 then an individual captured at two locations 9 days apart could be considered a movement in the network but two individuals captured 11 days apart could not)

mindate
The start date ("YYYY-MM-DD") of the study (i.e. when you want to build networks from)

maxdate
The end date ("YYYY-MM-DD") of the study (i.e. when you want to build networks until). Please provide as the day after the last day of the study.

netwindow
The period of time over which each network is built in months (i.e. netwindow=12 would correspond to yearly networks)

overlap
The amount of overlap between netwindows in months (i.e. overlap=2 would result in a second network window starting 2 months before the end of the first). Overlap=0 ensures no overlap between successive network windows

nextonly
(TRUE/FALSE). Determines whether a network edge is only created to the next capture of an individual or all captures within the intwindow. Defaults to FALSE

index
Defaults to FALSE. If FALSE edges are weighted by the number of movements. If TRUE then edges are weighted by the number of movements divided by the number of captures in a group

Value

A list of length 3 containing:

A list of edgelists (the same length as the number of network windows) containing the multiplex network for each of the netwindows as an array

a list of adjacency matrices (the same length as the number of ntwork windows) containing the multiplex network for each of the netwindows as an array

a matrix indicating which individuals occurred in each netwindow

Details

Multiplex networks connect locations that individuals have moved between within a particular interaction window with different layers being defined by the user. The time period for each network, together with the temporal and spatial restrictions on the capture window used to infer a movement can be defined by the user

Examples

if (FALSE) {
# load in data
data(cmrData2)

# set parameters
mindate<-"2010-01-01"
maxdate<-"2015-01-01"
intwindow<-60
netwindow<-12
overlap<-0

# create network
multimovenetdat<-MultiMoveNetCreate(data=cmrData2,
intwindow=intwindow,
mindate=mindate,
maxdate=maxdate,
netwindow=netwindow,
overlap=overlap,
nextonly=TRUE,
index=FALSE)
}