cmrSocPlot #' igraph plotting for CMRnet object

cmrSocPlot #' igraph plotting for CMRnet object

cmrSocPlot(
  nets,
  fixed_locs = c(TRUE, FALSE),
  locs = NULL,
  dynamic = c(TRUE, FALSE),
  rows = 3,
  ...
)

Arguments

nets
A list of igraph networks output by cmr_igraph

fixed_locs
(TRUE/FALSE). Whether to keep the the locations of the nodes fixed between network windows or allow them to change

locs
A matrix or dataframe giving x and y coordinates of all nodes in the full network. Can be used to provide user-defined layouts to the graph

dynamic
(TRUE/FALSE). If true this function will plot networks for the different network windows sequentially in the same plot window with a delay. If false then this function will plot the networks in a single, multipanel figure

rows
Sets the number of rows in the multipanel figure is dynamic=FALSE

...
Extra arguments from plot.igraph if desired

Value

A series of network plots showing the co-capture (social) network from each network window

Details

This function plots co-capture (social) networks produced by CMRnet. As well as the arguments detailed above, additional arguments from plot.igraph can be used to produce plots desired by the user

Examples

if (FALSE) {

# load in data
data(cmrData)

# set parameters
mindate<-"2010-01-01"
maxdate<-"2015-01-01"
intwindow<-60
netwindow<-12
overlap<-0
spacewindow <- 0

# Dynamic networks
netdat <- DynamicNetCreate(
 data = cmrData,
 intwindow = intwindow,
 mindate = mindate,
 maxdate = maxdate,
 netwindow = netwindow,
 overlap = overlap,
 spacewindow = spacewindow
)

# run permutations
networks<-cmr_igraph(netdat,type="social")
cmrSocPlot(nets=networks,fixed_locs=TRUE,dynamic=FALSE,rows=2,vertex.label=NA)
}